requirements.txt
# requirements.txt
requests
Pillow
